Handle Heroku's older version of git
Heroku has git 1.7.0, which doesn't support `blame -e`, which shows the
author's email address instead of their name. The name is harder to
parse, since it might contain spaces.

Longer term, we'll use `blame -p`, which will allow us to get both the
author's name and email address, but it's a more difficult format to
parse, and will take more time and effort.